Ref main structure
