+++ client/src/App.js
@@ -5,6 +5,7 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Main from "./pages/main/Main"
 import DataViz from "./pages/data_viz/DataViz";
 import PrivacyNotice from "./pages/privacy_notice/PrivacyNotice";
+import WebSites from "./pages/web_sites";
 
 import GlowEffect from "./components/GlowEffect";
 
@@ -14,8 +15,7 @@ const App = () => {
     <Router>
       <Routes>
         <Route path="/" element={<Main />}></Route>
-        <Route exact path="/visualizacion-de-datos" element={<DataViz />}></Route>
-        <Route exact path="/glow-effect" element={<GlowEffect />}></Route>
+        <Route exact path="/sitios-web" element={<WebSites />}></Route>
         <Route exact path="/aviso-de-privacidad" element={<PrivacyNotice />}></Route>
       </Routes>
     </Router>
