+++ client/src/components/inputs/Button.js
@@ -1,8 +1,11 @@
 import Button from "@mui/material/Button";
 
+
 const SecondaryButton = (props) => {
   return (
-    <Button variant="contained" color="primary" onClick={props.onClick}>
+    <Button variant="contained" color="primary" 
+    startIcon={props.icon}
+    onClick={props.onClick}>
       {props.children}
     </Button>
   );
