+++ client/src/components/layout/Footer.js
@@ -3,8 +3,14 @@ import TwitterIcon from '@mui/icons-material/Twitter';
 
 import classes from "./Footer.module.css";
 import { Link } from 'react-router-dom'
+import { Button } from '@mui/material';
 
 const Footer = (props) => {
+
+  const twitterHandler = () => {
+    window.open("https://twitter.com/alnitakmx", "_blank");
+  };
+  
   return (
     <footer className={classes.footer}>
       <div className={classes["left-side"]}>
@@ -17,12 +23,15 @@ const Footer = (props) => {
       <div className={classes["right-side"]}>
         <div className={classes["social-networks"]}>
           <ul>
-            <li><TwitterIcon /></li>
-       
+            <li>
+              <Button onClick={twitterHandler}>
+                <TwitterIcon />
+              </Button>  
+              </li>
           </ul>
         </div>
-        <span>55 5555 5555</span>
-        <span>|</span>
+        <span className={classes.item}>55 5555 5555</span>
+        <span className={classes.item}>|</span>
         
           
           <Tooltip title="¡Escríbenos!">
