+++ client/src/components/sections/get_in_touch/GetInTouch.js
@@ -7,7 +7,7 @@ import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
 import Button from "@mui/material/Button";
 import { useEffect, useState, useRef } from "react";
-import Paper from "../surfaces/Paper";
+import Paper from "../../surfaces/Paper";
 import emailjs from "@emailjs/browser";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
