+++ client/src/components/sections/solutions/Solutions.module.css
@@ -1,0 +1,34 @@
+.solutionsCtn {
+    font-family: "Roboto";
+    font-size: 18px;
+    display: flex;
+    justify-content: center;
+    margin-top: 120px;
+    padding-left: 1rem;
+    padding-right: 1rem;
+    line-height: 25px;
+}
+.solutionsBox {
+    margin-right: 1.5rem;
+    margin-left: 1.5rem;
+}
+.solutionsBox h3 {
+    text-align: left;
+    font-size: 20px;
+    font-weight: 700;
+    margin-bottom: 2.5rem;
+}
+
+/*Media Query for Mobile*/
+@media (max-width: 768px) {
+    .solutionsBox {
+        margin-top: 1rem;
+        width: 85%;
+    }
+    .solutionsCtn {
+        display: flex;
+        flex-direction: column;
+        height: auto;
+        margin-top: 0;
+    }  
+}
