+++ client/src/components/sections/us/Us.module.css
@@ -1,0 +1,89 @@
+.usCtn {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin-top: 175px;
+    margin-left: 1rem;
+    margin-right: 1rem;
+  }
+  .usBox {
+    width: 75%;
+  }
+  
+  .moto {
+    font-size: 24px;
+  }
+  .usCtn ul {
+    display: flex;
+    list-style: none;
+    align-items: center !important;
+    height: 20px;
+    margin: 0%;
+    padding: 0%;
+    margin-left: 1rem;
+    padding-bottom: 1rem;
+    padding-top: 1rem;
+  }
+  .usCtn li {
+    /* background-color: black; */
+    height: 135px;
+    width: 120px;
+    border-radius: 50%;
+    margin-right: 1rem;
+    margin-left: 1rem;
+    margin-bottom: 1rem;
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    align-items: center;
+  }
+  .usCtn li span {
+    font-size: 32px;
+  }
+  
+  .techDescr {
+    width: 300px;
+    background-color: rgba(0, 0, 0, 0.8);
+    padding: 25px;
+    border-radius: 30px 0 0 0;
+  }
+  
+  .usIcons {
+    height: 200px;
+    display: flex;
+    align-items: center;
+  }
+
+  /* Media Queries */
+  @media screen and (max-width: 768px) {
+    .usCtn ul {
+      display: block;
+    }
+    .usIcons {
+      height: 75vh;
+      display: flex;
+      align-items: initial;
+    }
+    .usCtn ul {
+      display: block;
+      list-style: none;
+      align-items: center !important;
+      height: 100vh;
+      margin: 0%;
+      padding: 0%;
+      margin-left: 1rem;
+      padding-bottom: 1rem;
+      padding-top: 1rem;
+    }
+    .usCtn {
+      margin-top: 2rem;
+    }
+    .usCtn h2 {
+      padding-left: 1rem;
+      padding-right: 1rem;
+    }
+    .usBox {
+      width: 100%;
+    }
+  }
