+++ client/src/components/snippets/three_models/SpaceShips.js
@@ -1,0 +1,49 @@
+{/* <Sky sunPosition={[1, .0005,1]} turbidity={0} rayleigh={rayleigh}/> */}
+            {/* <Model3
+              scale={0.03}
+              position={[getRandomArbitrary(-100, -50), 0, 0]}
+              rotation={[0, -Math.PI / 2, 0]}
+              xMov={0.05}
+              xRot={0.001}
+              yMov={1}
+            />
+            <Model3
+              scale={0.04}
+              position={[getRandomArbitrary(-100, -50), 15, 0]}
+              rotation={[0, -Math.PI / 2, 0]}
+              xMov={0.05}
+              xRot={0.005}
+              yMov={1}
+            />
+            <Model3
+              scale={0.06}
+              position={[getRandomArbitrary(-100, -50), -30, 0]}
+              rotation={[0, -Math.PI / 2, 0]}
+              xMov={0.05}
+              xRot={0.005}
+              yMov={1}
+            />
+            <Model4
+              scale={1.5}
+              position={[getRandomArbitrary(100, 50), 30, 0]}
+              rotation={[0, 0, 0]}
+              xMov={0.05}
+              xRot={0.00005}
+              yMov={1}
+            />
+            <Model4
+              scale={1.5}
+              position={[getRandomArbitrary(100, 50), -10, 0]}
+              rotation={[0, 0, 0]}
+              xMov={0.07}
+              xRot={0.00005}
+              yMov={1}
+            />
+            <Model5
+              scale={2}
+              position={[getRandomArbitrary(100, 150), -40, 0]}
+              rotation={[0, 0, 0]}
+              xMov={0.05}
+              xRot={0.00005}
+              yMov={0.025}
+            /> */}
