+++ client/src/pages/main/Main.module.css
@@ -2,10 +2,12 @@
 :root {
   font-family: "Roboto", sans-serif;
   font-size: 16px;
+  
 }
 
 .r3fCanvas {
   height: 100vh !important;
+  cursor: pointer !important;
 }
 
 /* .title {
@@ -28,7 +30,7 @@
   font-size: 18px;
   display: flex;
   justify-content: center;
-  margin-top: 150px;
+  margin-top: 120px;
   padding-left: 1rem;
   padding-right: 1rem;
   line-height: 25px;
@@ -104,12 +106,7 @@
   padding: 18px;
 }
 
-.getInTouchContainer {
-  position: absolute;
-  top: 150px;
-  left: 150px;
-  z-index: 1;
-}
+
 
 @media screen and (max-width: 768px) {
   .usCtn ul {
