+++ client/src/pages/web_sites/index.module.css
@@ -1,0 +1,14 @@
+.root {
+    margin: '2rem';
+    text-align: 'center';
+}
+.heading {
+    font-size: '2.5rem';
+    font-weight: 'bold';
+    color: 'black';
+  }
+.paragraph {
+    font-size: '1.25rem';
+    line-height: '2rem';
+    margin-bottom: '1rem';
+}
